Display different app grids depending on whether the user is logged in
