+++ frontend/src/components/AppGrid/FavoriteAppGrid/FavAppGrid.css
