+++ frontend/src/pages/apps.js
@@ -4,7 +4,7 @@ import "bootstrap/dist/css/bootstrap.min.css"
 
 import Layout from "../components/Layout/Layout"
 import SEO from "../components/Seo/Seo"
-import AppGrid from "../components/AppGrid/AppGrid"
+import AppGrid from "../components/AppGrid/PublicAppGrid"
 import GamesPage from "./apps/games"
 import ShoppingPage from "./apps/shopping"
 import ProductivityPage from "./apps/productivity"
