+++ frontend/src/pages/favorite-apps.js
@@ -3,13 +3,13 @@ import "bootstrap/dist/css/bootstrap.min.css"
 
 import Layout from "../components/Layout/Layout"
 import SEO from "../components/Seo/Seo"
-import FavoriteAppGrid from "../components/FavoriteAppGrid/FavoriteAppGrid"
+import MyFavoriteAppGrid from "../components/AppGrid/FavoriteAppGrid/MyFavoriteAppsGrid"
 
 const FavoriteAppsPage = () => (
   <Layout>
     <SEO title="Favorite Apps" />
     <h1>favorite apps</h1>
-    <FavoriteAppGrid></FavoriteAppGrid>
+    <MyFavoriteAppGrid></MyFavoriteAppGrid>
   </Layout>
 )
