+++ frontend/src/pages/manage-apps.js
@@ -5,7 +5,7 @@ import SEO from "../components/Seo/Seo"
 import SingleAppEdit from "./apps/single-app-edit"
 import { Router } from "@reach/router"
 import "bootstrap/dist/css/bootstrap.min.css"
-import AppGrid from "../components/AppGrid/AppGrid"
+import AppGrid from "../components/AppGrid/PublicAppGrid"
 import UserContext from "../shared/UserContext"
 import getMyApps from "../shared/fetchActions/getMyApps"
 import * as forFrontend from "../shared/convertForFrontend"
